Add nREPL port auto-discovery
The auto-discovery mechanism iterates over the root folders of the
projects you have open and checks them for .nrepl-port files. If it
finds more than one, it lets you choose from a list of ports it
discovered.

If it finds only one, it uses it as the default value in the port text
input. If it finds none, it uses the empty string as a default value.

It would be nice to have tests for all of this, but I'm not sure whether
there's a mechanism for checking input handlers.