Always operate on the sexp the cursor is skirting
Previously, if your cursor was next to a sexp rather than inside it, you
(as the user of the API) could choose whether you want to operate on the
form your cursor is skirting or the form it's inside (if any).

Since I couldn't think of a case where I wouldn't want to operate on the
form I'm skirting, I removed that choice. If the need arises, we can
restore it, hopefully with a better API.