Consolidate result and output views
I prefer having evaluation results, stdout, and stderr in the same view.
I separated them into different views previously because I couldn't
figure out a way to mix content with syntax highlighting (results) and
content without syntax highlighting (stdout/stderr).

The view.add_regions() method offers a way to do that. Unfortunately, it
is not possible to use view.add_regions() to only set the foreground
color. We therefore have to settle for the workaround described here:

https://github.com/sublimehq/sublime_text/issues/817#issuecomment-95211154

Currently, it is up to the user to add a scope called "repl.output" in
their color scheme and set the background color to a color that is
nearly but not exactly the same as the global background color. I'll
need to see if I can come up with a way to have Tutkain modify the color
scheme on the fly. I believe Rainbowth[1] does something like that.

[1]: https://github.com/whitequark/rainbowth