Make inline evaluation result a command argument
Turns out it's not a good idea to print inline evaluation result by
default. If the evaluation result is large, printing it inline is not
useful and hinders performance.

Instead, I opted to make it an argument to the tutkain_evaluate_form
command instead. That way users can add a key binding for evaluating
a form and showing inline result.