Improve keyword completion support
It turns out it makes more sense to only trigger auto-completion for
keywords if there's at least one non-colon character after the
colon(s).

I think, at least. It's quite tricky to get this right. For example,
after this commit, having a sublime-completions file with entries for
things like ##Inf and ##NaN doesn't work well any more. Not that that's
a critical issue.