Add compat mode for :require/:import indentation
This is the only concession I've made so far to the Tonsky formatting
rules. I don't mean to condone this style of indentation, but it is
simply so prevalent that handling indentation manually every time is too
much of a hassle.

I reserve the right to remove this heretical feature in the future.