Fix def and declare scopes
entity.name.variable.clojure is not included in the symbol list by
default. It is possible to use a .tmPreferences file to add it into the
symbol list, but then it won't get kind info.

Might consider restoring the previous scopes if this issue gets
resolved:

https://github.com/sublimehq/sublime_text/issues/3315

After all, a declare doesn't imply a function, and defs aren't
constants.