Improve performance of stdout printing
This is a somewhat evil hack that might come back to bite me in the ass,
but the performance gains are just so enticing. If this works, I'll need
to look into doing something similar with stderr. That's not as
critical, though, because folks don't typically print boatloads of
stderr.