Overhaul ClojureScript support
I've decided to target Shadow CLJS for now, because it appears that it
is not recommended to use the ClojureScript REPL APIs. I really wanted
to implement ClojureScript support in a tooling-agnostic manner, but
unfortunately, it does not appear to be possible. Or if it is, I haven't
figured out how to do it.