Optimize sexp.outermost()
Previously, sexp.outermost() would search until the end of the file for
the opening bracket. This obviously doesn't work great for namespaces
like clojure.core, which are pretty long.

As of this commit, we assume that the opening bracket is on the first
column. This means that cases where the opening bracket of a nested sexp
is on the first column will be handled incorrectly, but my hope is that
no one really does that. Proposals for a better solution are welcome.