Send local over as string instead of a symbol
For some reason, Tutkain would sometimes send a "locals" op for
an empty form (probably due to an error in the syntax definition). An
empty symbol is "valid" Clojure, but obviously doesn't print as valid,
so we'll send the form over as a string instead of a symbol and convert
it on the server.