Avoid output interleaving #86
This obviates tutkain.repl/pst, so removed that as well.

There's still no guarantee that stdout/stderr is printed before the eval
result on the client, but at least now there should be no interleaving.

This also greatly improves the performance of everything that prints and
flushes in a loop.